LPD-22076 - Follow up - Change test file extension and add non-breaking-space test coverage
